Fix integer overflow in postgres (#37)

+++ carmin-server/server/database/models/execution.py
@@ -2,7 +2,7 @@ import enum
 import uuid
 import time
 from flask_restful import fields
-from sqlalchemy import Column, String, Enum, Integer, ForeignKey
+from sqlalchemy import Column, String, Enum, Integer, BigInteger, ForeignKey
 from server import db
 from server.resources.models.execution import ExecutionStatus
 
@@ -39,7 +39,7 @@ class Execution(db.Model):
         error_code (int):
         start_date (int):
         end_date (int):
-        creator_username (str):        
+        creator_username (str):
     """
 
     identifier = Column(String, primary_key=True, default=execution_uuid)
@@ -49,9 +49,9 @@ class Execution(db.Model):
     status = Column(Enum(ExecutionStatus), nullable=False)
     study_identifier = Column(String)
     error_code = Column(Integer)
-    start_date = Column(Integer)
-    end_date = Column(Integer)
+    start_date = Column(BigInteger)
+    end_date = Column(BigInteger)
     creator_username = Column(
         String, ForeignKey("user.username"), nullable=False)
-    created_at = Column(Integer, default=current_milli_time)
-    last_update = Column(Integer, onupdate=current_milli_time)
+    created_at = Column(BigInteger, default=current_milli_time)
+    last_update = Column(BigInteger, onupdate=current_milli_time)
